refac: relocate test folder struct
