+++ src/test/java/org/example/kkumdoriland/acceptance/CucumberTest.java
@@ -1,9 +1,10 @@
-package org.example.kkumdoriland.cucumber;
+package org.example.kkumdoriland.acceptance;
 
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
 
 import io.cucumber.spring.CucumberContextConfiguration;
+import org.example.kkumdoriland.KkumdoriLandApplication;
 import org.junit.platform.suite.api.ConfigurationParameter;
 import org.junit.platform.suite.api.IncludeEngines;
 import org.junit.platform.suite.api.SelectClasspathResource;
@@ -18,7 +19,7 @@ import org.springframework.test.context.ActiveProfiles;
 @ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "org.example.kkumdoriland")
 @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty")
 @CucumberContextConfiguration
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes=KkumdoriLandApplication.class)
 public class CucumberTest {
 
 }
