+++ src/test/java/org/example/kkumdoriland/acceptance/steps/UserStepDef.java
@@ -1,7 +1,7 @@
-package org.example.kkumdoriland.cucumber.steps;
+package org.example.kkumdoriland.acceptance.steps;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.example.kkumdoriland.user.acceptance.UserSteps.유저_생성;
+import static org.example.kkumdoriland.acceptance.domains.UserSteps.유저_생성;
 import static org.example.kkumdoriland.utils.ResponseUtils.응답에서_id_조회;
 import static org.example.kkumdoriland.utils.ResponseUtils.응답의_STATUS_검증;
 
@@ -9,7 +9,7 @@ import io.cucumber.datatable.DataTable;
 import io.cucumber.java8.En;
 import java.util.Map;
 import java.util.List;
-import org.example.kkumdoriland.cucumber.AcceptanceContext;
+import org.example.kkumdoriland.acceptance.AcceptanceContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
