+++ src/test/java/org/example/kkumdoriland/integration/IntegrationTestBase.java
@@ -1,5 +1,6 @@
-package org.example.kkumdoriland.utils;
+package org.example.kkumdoriland.integration;
 
+import org.example.kkumdoriland.utils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -7,7 +8,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class IntegrationTest {
+public class IntegrationTestBase {
      @Autowired
      private DatabaseCleanup databaseCleanup;
